Migrate U256 and PieceDistance types to subspace-core-primitives.
