+++ crates/sp-consensus-subspace/src/verification.rs
@@ -25,10 +25,10 @@ use sp_api::HeaderT;
 use sp_consensus_slots::Slot;
 use sp_runtime::DigestItem;
 use subspace_archiving::archiver;
-use subspace_core_primitives::{PieceIndex, Randomness, Salt, Sha256Hash, Solution, Tag};
+use subspace_core_primitives::{PieceIndex, Randomness, Salt, Sha256Hash, Solution, Tag, U256};
 use subspace_solving::{
     derive_global_challenge, derive_target, is_tag_valid, verify_local_challenge,
-    verify_tag_signature, PieceDistance, SubspaceCodec,
+    verify_tag_signature, SubspaceCodec,
 };
 
 /// Errors encountered by the Subspace authorship task.
@@ -264,8 +264,8 @@ fn is_within_max_plot(
     if total_pieces < max_plot_size {
         return true;
     }
-    let max_distance_one_direction = PieceDistance::MAX / total_pieces * max_plot_size / 2;
-    PieceDistance::distance(&piece_index.into(), key.as_ref()) <= max_distance_one_direction
+    let max_distance_one_direction = U256::MAX / total_pieces * max_plot_size / 2;
+    U256::distance(&piece_index.into(), key.as_ref()) <= max_distance_one_direction
 }
 
 /// Parameters for checking piece validity
