+++ crates/subspace-core-primitives/Cargo.toml
@@ -18,6 +18,7 @@ parity-scale-codec = { version = "3.1.2", default-features = false, features = [
 scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
 serde = { version = "1.0.137", optional = true, features = ["derive"] }
 serde_arrays = "0.1.0"
+uint = { version = "0.9", default-features = false }
 
 # Ugly workaround for https://github.com/rust-lang/cargo/issues/1197
 [target.'cfg(any(target_os = "linux", target_os = "macos", all(target_os = "windows", target_env = "gnu")))'.dependencies.sha2]
@@ -39,4 +40,5 @@ std = [
     "scale-info/std",
     "serde",
     "sha2/std",
+    "uint/std",
 ]
