+++ crates/subspace-core-primitives/src/tests.rs
@@ -1,0 +1,6 @@
+use crate::U256;
+
+#[test]
+fn piece_distance_middle() {
+    assert_eq!(U256::MIDDLE, U256::MAX / 2);
+}
