+++ crates/subspace-farmer/src/dsn.rs
@@ -3,8 +3,9 @@ use futures::{Stream, StreamExt};
 use num_traits::WrappingAdd;
 use std::ops::{ControlFlow, Range};
 use std::sync::{Arc, Mutex};
-use subspace_core_primitives::{FlatPieces, PieceIndex, PieceIndexHash, PublicKey, Sha256Hash};
-use subspace_solving::U256;
+use subspace_core_primitives::{
+    FlatPieces, PieceIndex, PieceIndexHash, PublicKey, Sha256Hash, U256,
+};
 
 #[cfg(test)]
 mod tests;
