+++ crates/subspace-farmer/src/plot/tests.rs
@@ -1,8 +1,7 @@
-use crate::plot::Plot;
+use crate::plot::{PieceDistance, Plot};
 use rand::prelude::*;
 use std::sync::Arc;
 use subspace_core_primitives::{FlatPieces, Piece, PieceIndexHash, PIECE_SIZE};
-use subspace_solving::PieceDistance;
 use tempfile::TempDir;
 
 fn init() {
